Final CRUD USER dan ADMIN beserta middleware,api (Complete)

+++ app/Http/Controllers/PageController.php
@@ -6,14 +6,14 @@ use Auth;
 use Hash;
 use App\Models\Person;
 use Illuminate\Http\Request;
-
+use Illuminate\Http\Response;
 class PageController extends Controller
 {
 	public function getDashboard()
     {
         if(Auth::user()->role=='user')
         {
-            $users = Person::where('role','=','user')->get();
+           
             return view('master.adminPage', compact('users'));
         }
         else
@@ -30,8 +30,11 @@ class PageController extends Controller
     }
     public function editAdmin($id)
     {
-        $result= Person::where('username','=',$id)->first();
-        return view('admin.editadmin', compact('result'));
+        if(Auth::user()->role=='admin')
+         {
+            $result= Person::where('username','=',$id)->first();
+            return view('admin.editadmin', compact('result'));
+        }
     }
     public function createAdmin()
     {
@@ -54,19 +57,42 @@ class PageController extends Controller
         	$newadmin->email=$admin_email;
         	$admin_password=Hash::make($request->input('password'));
         	$newadmin->password=$admin_password;
-        	$newadmin->save();
+        	
         	$admin_username=$request->input('username');
         	$newadmin->username=$admin_username;
-        	$newadmin->save();
-  			return redirect()->route('dashboard')->withErrors('An new admin has just been added!');
-    }
+        	$unamecheck=$request->input('username');
+             $emailcheck=$request->input('email');
+             $unamequery = Person::where('username','=',$unamecheck)->first();
+            $emailquery = Person::where('email','=',$emailcheck)->first();
+            if(!is_null($unamequery)||!is_null($emailquery))
+            {   
+                return redirect()->route('createadmin')->withErrors('An admin with that username or email has already exists!');             
+            }
+             $newadmin->save();
+             return redirect()->route('dashboard')->withErrors('An new admin has just been added!'); 
+      }
+    
+
+    public function searchAdmin(Request $request)
+     {
+         $username=$request->get('keyword');
+         $admins = Person::where('username','like','%'.$username.'%')->where('role','=','admin')->get();
+         return view('admin.listadmin', compact('admins')); 
+     }
 
     public function saveAdmin(Request $request)
     {
         $keyword=$request->input('olduname');
         $admin_username=$request->input('username');
-        if($keyword!=$admin_username)
-       	{
+        if($keyword!=$admin_username){
+             $unamecheck=$request->input('username');
+             $emailcheck=$request->input('email');
+             $unamequery = Person::where('username','=',$unamecheck)->first();
+             $emailquery = Person::where('email','=',$emailcheck)->first();
+             if(!is_null($unamequery)||!is_null($emailquery))
+             {   
+                 return redirect()->route('editadmin',array('id'=>$keyword))->withErrors('An admin with that username or email has already exists!');             
+             }
        		$oldadmin = Person::where('username','=',$keyword)->first();
         	$oldadmin->delete();
         	$newadmin = new Person;
@@ -78,12 +104,15 @@ class PageController extends Controller
         	$newadmin->email=$admin_email;
         	$admin_password=Hash::make($request->input('password'));
         	$newadmin->password=$admin_password;
-        	$newadmin->save();
+        	
         	$newadmin->username=$admin_username;
         	$newadmin->save();
-       	}
-    	else if($keyword==$admin_username)
-    	{
+       } else if($keyword==$admin_username){
+             $emailcheck=$request->input('email');
+             $emailquery = Person::where('email','=',$emailcheck)->first();
+             if(!is_null($emailquery)){   
+                 return redirect()->route('editadmin',array('id'=>$keyword))->withErrors('An admin with that email has already exists!');             
+             }
     		$oldadmin = Person::where('username','=',$keyword)->first();
         	$admin_name=$request->input('nama');
         	$oldadmin->nama=$admin_name;
@@ -91,7 +120,7 @@ class PageController extends Controller
         	$oldadmin->email=$admin_email;
         	$admin_password=Hash::make($request->input('password'));
         	$oldadmin->password=$admin_password;
-        	$oldadmin->save();
+        	
         	$oldadmin->username=$admin_username;
         	$oldadmin->save();
       	}
@@ -103,8 +132,15 @@ class PageController extends Controller
 {
         $keyword=$request->input('olduname');
         $user_username=$request->input('username');
-        if($keyword!=$user_username)
-        {
+       if($keyword!=$user_username){
+             $unamecheck=$request->input('username');
+             $emailcheck=$request->input('email');
+             $unamequery = Person::where('username','=',$unamecheck)->first();
+             $emailquery = Person::where('email','=',$emailcheck)->first();
+             if(!is_null($unamequery)||!is_null($emailquery))
+             {   
+                 return redirect()->route('edituser',array('id'=>$keyword))->withErrors('An user with that username or email has already exists!');             
+            }
             $olduser = Person::where('username','=',$keyword)->first();
             $olduser->delete();
             $newuser = new Person;
@@ -116,12 +152,16 @@ class PageController extends Controller
             $newuser->email=$user_email;
             $user_password=Hash::make($request->input('password'));
             $newuser->password=$user_password;
-            $newuser->save();
+            
             $newuser->username=$user_username;
             $newuser->save();
         }
-        else if($keyword==$user_username)
-        {
+         else if($keyword==$user_username){
+             $emailcheck=$request->input('email');
+             $emailquery = Person::where('email','=',$emailcheck)->first();
+             if(!is_null($emailquery)){   
+                 return redirect()->route('edituser',array('id'=>$keyword))->withErrors('An user with that email has already exists!');             
+             }
             $olduser = Person::where('username','=',$keyword)->first();
             $user_name=$request->input('nama');
             $olduser->nama=$user_name;
@@ -129,7 +169,6 @@ class PageController extends Controller
             $olduser->email=$user_email;
             $user_password=Hash::make($request->input('password'));
             $olduser->password=$user_password;
-            $olduser->save();
             $olduser->username=$user_username;
             $olduser->save();
         }
@@ -138,8 +177,11 @@ class PageController extends Controller
 
      public function editUser($id)
     {
-        $result= Person::where('username','=',$id)->first();
-        return view('user.edituser', compact('result'));
+        if(Auth::user()->role=='admin')
+         {
+             $result= Person::where('username','=',$id)->first();
+             return view('user.edituser', compact('result'));
+         }
     }
     public function createUser()
     {
@@ -151,6 +193,13 @@ class PageController extends Controller
         $result->delete();
         return redirect()->route('dashboard')->withErrors('An new user has just been deleted!');
 }
+
+public function searchUser(Request $request)
+     {
+         $username=$request->get('keyword');
+         $users = Person::where('username','like','%'.$username.'%')->where('role','=','user')->get();
+         return view('user.listuser', compact('users')); 
+     }
     public function registerUser(Request $request)
     {
             $newuser = new Person;
@@ -162,11 +211,19 @@ class PageController extends Controller
             $newuser->email=$user_email;
             $user_password=Hash::make($request->input('password'));
             $newuser->password=$user_password;
-            $newuser->save();
+            
             $user_username=$request->input('username');
             $newuser->username=$user_username;
-            $newuser->save();
-            return redirect()->route('dashboard')->withErrors('An new admin has just been added!');
+            $unamecheck=$request->input('username');
+             $emailcheck=$request->input('email');
+             $unamequery = Person::where('username','=',$unamecheck)->first();
+             $emailquery = Person::where('email','=',$emailcheck)->first();
+             if(!is_null($unamequery)||!is_null($emailquery))
+             {   
+                 return redirect()->route('createuser')->withErrors('An admin with that username or email has already exists!');             
+             }
+              $newadmin->save();
+              return redirect()->route('dashboard')->withErrors('An new User has just been added!'); 
     }
 }
 ?>
